DS_Assign_1/requirements.txt
djangorestframework
psycopg2